requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
flake8==6.0.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mongoengine==0.27.0
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.3.3
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==2.3.4
WTForms==3.0.1
zipp==3.15.0
